requirements.txt
FLASK==3.0.3
FLASK-SQLALCHEMY==3.1.1
FLASK-WTF==1.2.1
FLASK-LOGIN==0.6.3
EMAIL-VALIDATOR==2.1.1
FLASK-BOOTSTRAP==3.3.7.1